+++ src/deal-types.js
