+++ src/display-winner.js
